+++ sdk/keyvault/keyvault-admin/src/accessControlClient.ts
@@ -7,7 +7,8 @@ import {
   isTokenCredential,
   signingPolicy,
   createPipelineFromOptions,
-  InternalPipelineOptions
+  InternalPipelineOptions,
+  RestResponse
 } from "@azure/core-http";
 import { PagedAsyncIterableIterator } from "@azure/core-paging";
 
@@ -177,10 +178,10 @@ export class KeyVaultAccessControlClient {
     roleScope: KeyVaultRoleScope,
     name: string,
     options: DeleteRoleAssignmentOptions = {}
-  ): Promise<void> {
-    return withTrace("deleteRoleAssignment", options, async (updatedOptions) => {
-      await this.client.roleAssignments.delete(this.vaultUrl, roleScope, name, updatedOptions);
-    });
+  ): Promise<RestResponse> {
+    return withTrace("deleteRoleAssignment", options, (updatedOptions) =>
+      this.client.roleAssignments.delete(this.vaultUrl, roleScope, name, updatedOptions)
+    );
   }
 
   /**
@@ -486,9 +487,9 @@ export class KeyVaultAccessControlClient {
     roleScope: KeyVaultRoleScope,
     name: string,
     options: DeleteRoleDefinitionOptions = {}
-  ): Promise<void> {
-    return withTrace("deleteRoleDefinition", options, async (updatedOptions) => {
-      await this.client.roleDefinitions.delete(this.vaultUrl, roleScope, name, updatedOptions);
-    });
+  ): Promise<RestResponse> {
+    return withTrace("deleteRoleDefinition", options, (updatedOptions) =>
+      this.client.roleDefinitions.delete(this.vaultUrl, roleScope, name, updatedOptions)
+    );
   }
 }